GeoFormatTypes.jl

GeoFormatTypes.GeoFormatTypes — Module

GeoFormatTypes

GeoFormatTypes defines wrapper types to make it easy to pass Geographic formats like WellKnownText or GeoJSON between packages while keeping information about what format is contained - instead of passing a String or Int that could be from any type.

Wrapper types also allow methods such as convert to work with data in multiple formats, instead of defining lists of format-specific handling methods.

GeoFormatTypes.AbstractWellKnownText — Type

Well known text has a number of versions and standards, and can represent coordinate reference systems or geometric data.

GeoFormatTypes.CoordSys — Type

Mapinfo CoordSys string

GeoFormatTypes.CoordinateReferenceSystemFormat — Type

Formats representing coordinate reference systems

GeoFormatTypes.EPSG — Type

EPSG code representing a coordinate reference system from the EPSG spatial reference system registry.

GeoFormatTypes.EPSG — Method

Constructor for "EPSG:1234" string input

GeoFormatTypes.ESRIWellKnownText — Type

Well known text following the ESRI standard

GeoFormatTypes.GML — Type

Geography Markup Language

GeoFormatTypes.GeoFormat — Type

Abstract supertype for geospatial data formats

GeoFormatTypes.GeoJSON — Type

GeoJSON String or Dict

GeoFormatTypes.GeometryFormat — Type

Formats representing geometries. These wrappers simply mark string formats that may optionally be converted to Geoetry objects at a later point.

GeoFormatTypes.KML — Type

Keyhole Markup Language

GeoFormatTypes.MixedFormat — Type

Formats that may hold either or both coordinate reference systems and geometries.

GeoFormatTypes.ProjString — Type

Proj string

GeoFormatTypes.WellKnownBinary — Type

Well known binary

GeoFormatTypes.WellKnownText — Type

Well known text v1 following the OGC standard

GeoFormatTypes.WellKnownText2 — Type

Well known text v2 following the new OGC standard

GeoFormatTypes.val — Function

val(f::GeoFormat)

Get the contained value of a GeoFormat type.